extract OutputInfoCollector from extension.ts
