+++ src/outputInfoCollector.ts
@@ -1,0 +1,43 @@
+import { OutputChannel, window, ViewColumn } from "vscode";
+import { logger } from "./log";
+
+export class OutputInfoCollector implements OutputChannel {
+	private channel: OutputChannel = null;
+
+	constructor(public name: string) {
+		this.channel = window.createOutputChannel(this.name);
+	}
+
+	append(value: string): void {
+		logger.info(value);
+		this.channel.append(value);
+	}
+
+	appendLine(value: string): void {
+		logger.info(value);
+		this.channel.appendLine(value);
+	}
+
+	replace(value: string): void {
+		this.clear();
+		this.append(value);
+	}
+
+	clear(): void {
+		this.channel.clear();
+	}
+
+	show(preserveFocus?: boolean): void;
+	show(column?: ViewColumn, preserveFocus?: boolean): void;
+	show(column?: any, preserveFocus?: any) {
+		this.channel.show(column, preserveFocus);
+	}
+
+	hide(): void {
+		this.channel.hide();
+	}
+
+	dispose(): void {
+		this.channel.dispose();
+	}
+}
