+++ src/syntaxLanguageClient.ts
@@ -4,9 +4,10 @@ import * as net from "net";
 import { DidChangeConfigurationNotification, LanguageClientOptions } from "vscode-languageclient";
 import { LanguageClient, ServerOptions, StreamInfo } from "vscode-languageclient/node";
 import { apiManager } from "./apiManager";
-import { ClientErrorHandler, getJavaConfig, OutputInfoCollector } from "./extension";
+import { ClientErrorHandler, getJavaConfig } from "./extension";
 import { ClientStatus, ExtensionAPI } from "./extension.api";
 import { logger } from "./log";
+import { OutputInfoCollector } from "./outputInfoCollector";
 import { StatusNotification } from "./protocol";
 import { ServerMode } from "./settings";
 import { snippetCompletionProvider } from "./snippetCompletionProvider";
